Show results if already scanned
If the Dragonfly API returns a 409, then make another request to check
if the package has results already. If it does, then simply display
those results. If it doesn't, then that means the package is currently
in a "queued" or "pending" state. We can just send a feedback message
saying this in this case.