Improve queued package status check
Instead of creating a new variable `package` assigned to the first item
of `scan_results` (or `None` if its' empty), we simply check if the
`scan_results` list is not empty - if it's not, use the first item. If
it is, display a message to the user.

The `package` variable obtained from getting the first item of
`scan_results` was unused regardless, so this only made the code harder
to read.